ModulesDatabase in kegg.py -> metabolism/modulesdb.py
