+++ anvio/reactionnetwork.py
@@ -29,7 +29,6 @@ from argparse import Namespace
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Set, Tuple, Union, Iterable
 
-import anvio.kegg as kegg
 import anvio.utils as utils
 import anvio.dbinfo as dbinfo
 import anvio.tables as tables
@@ -42,6 +41,7 @@ from anvio import DEBUG, __file__ as ANVIO_PATH, __version__ as VERSION
 from anvio.dbops import ContigsDatabase, PanDatabase, PanSuperclass, ProfileDatabase
 
 from anvio.metabolism.context import KeggContext
+from anvio.metabolism.modulesdb import ModulesDatabase
 from anvio.metabolism.downloads import _download_worker
 
 __author__ = "Developers of anvi'o (see AUTHORS.txt)"
@@ -5477,7 +5477,7 @@ class KEGGData:
     kegg_context : anvio.metabolism.context.KeggContext
         This contains anvi'o KEGG database attributes, such as filepaths.
 
-    modules_db : anvio.kegg.ModulesDatabase
+    modules_db : anvio.metabolism.modulesdb.ModulesDatabase
         The anvi'o modules database from which KEGG data is loaded.
 
     modules_db_hash : str
@@ -5560,9 +5560,7 @@ class KEGGData:
 
         utils.is_kegg_modules_db(self.kegg_context.kegg_modules_db_path)
 
-        self.modules_db = kegg.ModulesDatabase(
-            self.kegg_context.kegg_modules_db_path, argparse.Namespace(quiet=True)
-        )
+        self.modules_db = ModulesDatabase(self.kegg_context.kegg_modules_db_path, argparse.Namespace(quiet=True))
         self.modules_db_hash = self.modules_db.db.get_meta_value('hash')
 
         self.ko_data: Dict[str, Dict[str, Any]] = {}
